arreglos generales

+++ application/models/Categorias_model.php
@@ -13,37 +13,48 @@ class Categorias_model extends CI_Model {
     	$query = $this->db->get('categorias');
     	return $query->result();
     }	
-	function editar($id = NULL, $atributo = NULL, $valor = NULL){
-        if($id == NULL OR $atributo == NULL OR $valor == NULL){
-            return array('res'=>'bad','msj'=>'Error en la edición.'); }
-        $this->db->trans_start();
-        $object = array($atributo => $valor);
-        $this->db->where('id', $id);
-        $this->db->update('categorias', $object);
-        $this->db->trans_complete();
-        if ($this->db->trans_status() === FALSE) {
-            return array('res'=>'bad','msj'=>'Error en la edición.'); }
-        else{ return array('res'=>'ok'); }
-            
+	
+    function editar($id = NULL, $atributo = NULL, $valor = NULL){
+        if($id != NULL AND $atributo != NULL AND $valor != NULL){
+            if($atributo =="nombre"){
+                $this->db->where('nombre', $valor);
+                if($this->db->count_all_results('categorias')>0){
+                    return array('res'=>'bad','msj'=>'ERROR en edición. Ya existe una categoria con ese nombre.'); }
+            }
+            $this->db->trans_start();
+            $object = array($atributo => $valor);
+            $this->db->where('id', $id);
+            $this->db->update('categorias', $object);
+            $this->db->trans_complete();
+
+            if ($this->db->trans_status() === FALSE) {
+                return array('res'=>'bad','msj'=>'ERROR en edición.'); }
+            else {return array('res'=>'ok'); }
+        }
     }
+    
     function crear($nombre = null, $descripcion = null){
         if($nombre == NULL OR $descripcion == null){
-            return array('res'=>'bad','msj'=>'Error en la creación.'); }
+            return array('res'=>'bad','msj'=>'ERROR en creación.'); }
 
         if(strlen($nombre)<3){
-            return array('res'=>'bad','msj'=>'Ingresa un nombre adecuado'); }
+            return array('res'=>'bad','msj'=>'ERROR, Ingresa un nombre adecuado'); }
 
         $this->db->where('nombre', $nombre);
         if($this->db->count_all_results('categorias')>0){
-            return array('res'=>'bad','msj'=>'Ya existe una categoría con este nombre'); }
+            return array('res'=>'bad','msj'=>'ERROR, Ya existe una categoría con este nombre'); }
 
         $object = array('nombre' => $nombre, 'descripcion' => $descripcion);
         $this->db->insert('categorias', $object);
         return array('res'=>'ok','id'=>$this->db->insert_id());
     }
+    
     function eliminar($id = null){
         if($id == null){
-            return array('res'=>'bad','msj'=>'Error en la inserción.'); }
+            return array('res'=>'bad','msj'=>'ERROR en inserción.'); }
+        $this->db->where('idcategoria', $id);
+        if($this->db->count_all_results('pro_cat')>0){
+            return array('res'=>'bad','msj'=>'ERROR en inserción. Hay productos asociados a esta categoria.'); }
         $this->db->where('id', $id);
         $this->db->delete('categorias');
         return array('res'=>'ok');
