+++ application/models/Marcas_model.php
@@ -12,8 +12,14 @@ class Marcas_model extends CI_Model {
     	$query = $this->db->get('marcas');
     	return $query->result();
     }
+    
     function editar($id = NULL, $atributo = NULL, $valor = NULL){
         if($id != NULL AND $atributo != NULL AND $valor != NULL){
+            if($atributo =="nombre"){
+                $this->db->where('nombre', $valor);
+                if($this->db->count_all_results('marcas')>0){
+                    return array('res'=>'bad','msj'=>'ERROR en edición. Ya existe una Marca con ese nombre.'); }
+            }
             $this->db->trans_start();
             $object = array($atributo => $valor);
             $this->db->where('id', $id);
@@ -21,28 +27,33 @@ class Marcas_model extends CI_Model {
             $this->db->trans_complete();
 
             if ($this->db->trans_status() === FALSE) {
-                return array('res'=>'bad','msj'=>'Error en la edición.'); }
+                return array('res'=>'bad','msj'=>'ERROR en edición.'); }
             else {return array('res'=>'ok'); }
         }
     }
+    
     function crear($nombre = null, $descripcion = null){
         if($nombre == NULL OR $descripcion == null){
-            return array('res'=>'bad','msj'=>'Error en la creación.'); }
+            return array('res'=>'bad','msj'=>'ERROR en creación.'); }
 
         if(strlen($nombre)<3){
-            return array('res'=>'bad','msj'=>'Ingresa un nombre adecuado'); }
+            return array('res'=>'bad','msj'=>'ERROR, Ingresa un nombre adecuado'); }
 
         $this->db->where('nombre', $nombre);
         if($this->db->count_all_results('marcas')>0){
-            return array('res'=>'bad','msj'=>'Ya existe una caracteristica con este nombre'); }
+            return array('res'=>'bad','msj'=>'ERROR, Ya existe una Marca con este nombre'); }
 
         $object = array('nombre' => $nombre, 'descripcion' => $descripcion);
         $this->db->insert('marcas', $object);
         return array('res'=>'ok','id'=>$this->db->insert_id());
     }
+    
     function eliminar($id = null){
         if($id == null){
-            return array('res'=>'bad','msj'=>'Error en la inserción.'); }
+            return array('res'=>'bad','msj'=>'ERROR en eliminación.'); }
+        $this->db->where('idmarca', $id);
+        if($this->db->count_all_results('productos')>0){
+            return array('res'=>'bad','msj'=>'ERROR no se puede eliminar. Hay productos asociados a esta Marca.'); }
         $this->db->where('id', $id);
         $this->db->delete('marcas');
         return array('res'=>'ok');
